Downgrade jax to 0.4.23 in biomakerca notebooks.
This is because we are using evojax which needs an old version of flax. Old flax is not compatible with new jax ([see this discussion](https://github.com/google/flax/issues/3180)).

This is a temporary patch. Hopefully. We will figure out how to either fix evojax or stop using it for future experiments.

Also, fix a bug on the sexual reproduction colab where I was importing a function that I deleted, causing the colab to fail.

PiperOrigin-RevId: 638631920